requirements.txt
asgiref==3.8.1
Django==5.0.3
django-chartjs==2.3.0
django-guardian==2.4.0
djangorestframework==3.15.1
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
